requirements.txt
asgiref==3.5.2
Django==4.1.2
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.5
sqlparse==0.4.3
whitenoise==6.2.0
Pillow==9.3.0
psycopg2-binary==2.9.5